Changed models.py file.
Created User class which is related to registration, last line it's not the column it's the relationship with products,
that user has bought. The owner of the product displays in SalableGood model as the last column with id as foreign key to User.
By default the user has 1000$ in his|her account.